CRUD has been developed for UserController

+++ src/main/java/com/github/gad1k/WebBI/entity/UserRepository.java
@@ -2,5 +2,5 @@ package com.github.gad1k.WebBI.entity;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<DatabaseUser, Integer> {
+public interface UserRepository extends CrudRepository<DatabaseUser, Long> {
 }
